修改shopping order id为Long类型,测试抓取order存入数据库

+++ src/main/java/ltd/yuhan/erp/mapper/ShoppingOrderMapper.java
@@ -6,11 +6,11 @@ import org.apache.ibatis.annotations.Mapper;
 
 @Mapper
 public interface ShoppingOrderMapper {
-    int deleteByPrimaryKey(Integer id);
+    int deleteByPrimaryKey(Long id);
 
     int insert(ShoppingOrder record);
 
-    ShoppingOrder selectByPrimaryKey(Integer id);
+    ShoppingOrder selectByPrimaryKey(Long id);
 
     List<ShoppingOrder> selectAll();
